Moved components folder out of app directory
