+++ app/error.jsx
@@ -3,7 +3,7 @@
 import spanishdex from "@/public/logo.svg";
 import Image from "next/image";
 import Link from "next/link";
-import HomepageSection from "./components/HomepageSection";
+import HomepageSection from "@/components/HomepageSection";
 import Button from "react-bootstrap/Button";
 import rightArrow from "@/public/icons/arrow_right.svg";
 import downArrow from "@/public/icons/arrow_down.svg";
