+++ components/headers/ProfilePictureToggler.jsx
