+++ components/modals/LogInModal.jsx
@@ -5,12 +5,12 @@ import Stack from "react-bootstrap/Stack"
 import Form from "react-bootstrap/Form";
 import Container from "react-bootstrap/Container"
 import Alert from "react-bootstrap/Alert";
-import IconButton from "../IconButton";
+import IconButton from "@/components/IconButton";
 import { useEffect, useState } from "react";
 import { signIn } from 'next-auth/react';
 import { useRouter } from "next/navigation";
-import GoogleAuthButton from "../GoogleAuthButton";
-import OrSeparator from "../OrSeparator";
+import GoogleAuthButton from "@/components/GoogleAuthButton";
+import OrSeparator from "@/components/OrSeparator";
 
 function LogInModal(props) {
   const router = useRouter();
